Add files via upload
Each of these zip files are different sections of a larger folder created using pyinstaller to make the application and executable file. Each one needs to be downloaded and extracted into a single folder to allow use. The application zip folder contains 'KRDPyCharm.exe' which is the main application file to be ran. Additional changes will be made and then uploaded after user tests to provide clarification.