Fixed Bug with concentration calculation
In PFR and Membrane reactor concentration is calculated with C_n = cao*(Theta(n)+Nu(n)*x)*(To/T)*y/(1+eps*x) however the (To/T)*y/(1+eps*x) section is only for gas phase reactions so a condition of reaction type gas must be met to include this in the calculation. To remedy this section is raised to the conditional test, if true = 1 and remains if false = 0 and becomes multiplication of 1. Same may be said for multi reaction calculations.